# src/swagger.yaml
openapi: 3.0.3
info:
  title: NetScopeNMS REST API
  description: |
    AI 기반 네트워크 관리 시스템 (Network Management System) API 문서
    
    ## 주요 기능
    - **장비 관리**: 네트워크 장비 CRUD, SNMP 연결 테스트, 인터페이스 검색
    - **메트릭 수집**: 실시간 및 이력 데이터 조회 (CPU, 메모리, 트래픽 등)
    - **알람 관리**: 임계값 기반 알람 생성 및 관리
    - **AI 분석**: OpenAI 기반 근본 원인 분석, 예측, 리포트 생성
    - **사용자 관리**: JWT 기반 인증, 역할 기반 권한 관리
    
    ## 인증
    Bearer Token (JWT) 방식을 사용합니다.
    ```
    Authorization: Bearer <token>
    ```
    
    ## 오류 응답 형식
    ```json
    {
      "success": false,
      "error": "Error type",
      "message": "상세 메시지",
      "details": [] // 선택적 상세 정보
    }
    ```
    
  version: 1.0.0
  contact:
    name: NetScopeNMS Team
    email: support@netscopenms.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.netscopenms.com/api/v1
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /devices:
    get:
      tags:
        - Devices
      summary: 장비 목록 조회
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            type: string
            enum: [up, down, warning, unknown, all]
        - name: type
          in: query
          schema:
            type: string
            enum: [router, switch, server, firewall, access_point, other, all]
        - name: search
          in: query
          schema:
            type: string
          description: 이름, IP, 위치 검색
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, ip_address, status, device_type, created_at, last_poll_time]
            default: name
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
            default: ASC
      responses:
        '200':
          description: 장비 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      devices:
                        type: array
                        items:
                          $ref: '#/components/schemas/Device'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
    post:
      tags:
        - Devices
      summary: 장비 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
      responses:
        '201':
          description: 등록 성공
        '409':
          description: IP 주소 중복
...
  /devices/test-connection:
    post:
      tags:
        - Devices
      summary: 신규 장비 SNMP 연결 테스트
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnmpTestRequest'
      responses:
        '200':
          description: 테스트 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnmpTestResponse'
  /devices/{id}:
    get:
      tags:
        - Devices
      summary: 장비 상세 조회
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: 장비 상세 정보
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      device:
                        $ref: '#/components/schemas/DeviceDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Devices
      summary: 장비 수정
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdate'
      responses:
        '200':
          description: 수정 성공
    delete:
      tags:
        - Devices
      summary: 장비 삭제
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: 삭제 성공
  /devices/{id}/test-connection:
    post:
      tags:
        - Devices
      summary: 기존 장비 SNMP 연결 테스트
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: 테스트 결과
  /devices/{id}/discover-interfaces:
    post:
      tags:
        - Devices
      summary: 인터페이스 검색
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: 검색된 인터페이스 목록
  /devices/{id}/interfaces:
    get:
      tags:
        - Devices
      summary: 장비 인터페이스 목록
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - name: status
          in: query
          schema:
            type: string
            enum: [up, down, testing, unknown, dormant]
      responses:
        '200':
          description: 인터페이스 목록
  /devices/{id}/toggle:
    patch:
      tags:
        - Devices
      summary: 장비 활성화/비활성화 토글
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: 토글 성공
  /devices/{id}/poll:
    post:
      tags:
        - Devices
      summary: 수동 폴링 실행
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: 폴링 결과

  # ==================== 메트릭 ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT 토큰을 사용한 인증
        
        로그인 후 받은 토큰을 `Authorization: Bearer <token>` 형식으로 전송
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 페이지 번호
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 페이지당 항목 수
    DeviceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: 장비 ID
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: 전체 항목 수
        page:
          type: integer
          description: 현재 페이지
        limit:
          type: integer
          description: 페이지당 항목 수
        totalPages:
          type: integer
          description: 전체 페이지 수
...
    Device:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        ip_address:
          type: string
          format: ipv4
        device_type:
          type: string
          enum: [router, switch, server, firewall, access_point, other]
        vendor:
          type: string
        model:
          type: string
        location:
          type: string
        status:
          type: string
          enum: [up, down, warning, unknown]
        is_enabled:
          type: boolean
        snmp_version:
          type: string
          enum: ['1', '2c', '3']
        poll_interval:
          type: integer
        last_poll_time:
          type: string
          format: date-time
        interfaceCount:
          type: integer
    DeviceDetail:
      allOf:
        - $ref: '#/components/schemas/Device'
        - type: object
          properties:
            description:
              type: string
            snmp_port:
              type: integer
            sys_descr:
              type: string
            sys_name:
              type: string
            sys_uptime:
              type: integer
            uptimeFormatted:
              type: string
            interfaces:
              type: array
              items:
                $ref: '#/components/schemas/Interface'
            credentials:
              type: object
    DeviceCreate:
      type: object
      required:
        - name
        - ip_address
      properties:
        name:
          type: string
          maxLength: 100
          example: Core-Router-01
        ip_address:
          type: string
          format: ipv4
          example: 192.168.1.1
        device_type:
          type: string
          enum: [router, switch, server, firewall, access_point, other]
          default: other
        vendor:
          type: string
          example: Cisco
        model:
          type: string
          example: ISR 4331
        location:
          type: string
          example: 본관 서버실
        description:
          type: string
        snmp_version:
          type: string
          enum: ['1', '2c', '3']
          default: '2c'
        snmp_port:
          type: integer
          default: 161
        poll_interval:
          type: integer
          minimum: 30
          maximum: 3600
          default: 60
        community:
          type: string
          description: SNMPv1/v2c 커뮤니티 문자열
          example: public
        username:
          type: string
          description: SNMPv3 사용자명
        security_level:
          type: string
          enum: [noAuthNoPriv, authNoPriv, authPriv]
          description: SNMPv3 보안 레벨
        auth_protocol:
          type: string
          enum: [MD5, SHA, SHA-256, SHA-384, SHA-512]
        auth_password:
          type: string
        priv_protocol:
          type: string
          enum: [DES, AES, AES-128, AES-192, AES-256]
        priv_password:
          type: string
    DeviceUpdate:
      type: object
      properties:
        name:
          type: string
        ip_address:
          type: string
          format: ipv4
        device_type:
          type: string
          enum: [router, switch, server, firewall, access_point, other]
        vendor:
          type: string
        model:
          type: string
        location:
          type: string
        description:
          type: string
        snmp_version:
          type: string
          enum: ['1', '2c', '3']
        snmp_port:
          type: integer
        poll_interval:
          type: integer
        is_enabled:
          type: boolean
    Interface:
      type: object
      properties:
        id:
          type: integer
        if_index:
          type: integer
        if_name:
          type: string
        if_descr:
          type: string
        if_type:
          type: integer
        if_speed:
          type: integer
          description: 인터페이스 속도 (bps)
        if_status:
          type: string
          enum: [up, down, testing, unknown, dormant, notPresent, lowerLayerDown]
        if_admin_status:
          type: string
          enum: [up, down, testing]
    SnmpTestRequest:
      type: object
      required:
        - ip_address
      properties:
        ip_address:
          type: string
          format: ipv4
        snmp_version:
          type: string
          enum: ['1', '2c', '3']
          default: '2c'
        snmp_port:
          type: integer
          default: 161
        community:
          type: string
          default: public
        username:
          type: string
        security_level:
          type: string
        auth_protocol:
          type: string
        auth_password:
          type: string
        priv_protocol:
          type: string
        priv_password:
          type: string
    SnmpTestResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            success:
              type: boolean
            responseTimeMs:
              type: integer
            sysDescr:
              type: string
            sysName:
              type: string
            sysUpTime:
              type: integer
            uptimeFormatted:
              type: string
            detectedVendor:
              type: string
            detectedType:
              type: string
            error:
              type: string
...
  responses:
...
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Not Found
            message: 요청한 리소스를 찾을 수 없습니다
...